requirements.txt
Django==4.2.23
python-decouple==3.8
openai==1.35.13
python-dotenv==1.0.1
Pillow==10.4.0
PyPDF2==3.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-magic==0.4.27
python-docx==0.8.11
